@@ -1,35 +1,20 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.BlazorWebAssembly">
+﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <!--Temporal hack that fixes compilation in VS-->
-    <TypeScriptCompileBlocked>true</TypeScriptCompileBlocked>
-    <EmccTotalMemory>16777216</EmccTotalMemory>
+	<TargetFramework>net7.0</TargetFramework>
+	<RuntimeIdentifier>browser-wasm</RuntimeIdentifier>
+	<WasmMainJSPath>AppBundle\main.js</WasmMainJSPath>
+	<OutputType>Exe</OutputType>
   </PropertyGroup>
 
-  <!-- In debug, make builds faster by reducing optimizations -->
-  <PropertyGroup Condition="'$(Configuration)' == 'Debug'">
-    <WasmNativeStrip>false</WasmNativeStrip>
-    <EmccCompileOptimizationFlag>-O1</EmccCompileOptimizationFlag>
-    <RunAOTCompilation>false</RunAOTCompilation>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-    <Optimize>true</Optimize>
-    <WasmNativeStrip>true</WasmNativeStrip>
-	<EmccCompileOptimizationFlag>-O3</EmccCompileOptimizationFlag>
-    <EmccLinkOptimizationFlag>-O3</EmccLinkOptimizationFlag>
-    <RunAOTCompilation>true</RunAOTCompilation>
-  </PropertyGroup>
-  
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="6.0.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.DevServer" Version="6.0.0" PrivateAssets="all" />
-    <PackageReference Include="Avalonia.Web.Blazor" Version="$(AvaloniaVersion)" />
+    <WasmExtraFilesToDeploy Include="AppBundle\**" />
+  </ItemGroup>
+	
+  <ItemGroup>
+    <PackageReference Include="Avalonia.Web" Version="$(AvaloniaVersion)" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\AvaloniaTest\AvaloniaTest.csproj" />
   </ItemGroup>
-
 </Project>
