@@ -1,14 +1,14 @@
 ﻿#if (!ReactiveUIToolkitChosen)
 using CommunityToolkit.Mvvm.ComponentModel;
 #endif
-namespace AvaloniaAppTemplate.ViewModels
-{
+
+namespace AvaloniaAppTemplate.ViewModels;
+
 #if (!ReactiveUIToolkitChosen)
-    public class MainWindowViewModel : ObservableObject
+public class MainWindowViewModel : ObservableObject
 #else
-    public class MainWindowViewModel : ViewModelBase
+public class MainWindowViewModel : ViewModelBase
 #endif
-    {
-        public string Greeting => "Welcome to Avalonia!";
-    }
+{
+    public string Greeting => "Welcome to Avalonia!";
 }
