@@ -2,13 +2,12 @@ using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace AvaloniaAppTemplate.Namespace
+namespace AvaloniaAppTemplate.Namespace;
+
+public partial class NewWindow : Window
 {
-    public partial class NewWindow : Window
+    public NewWindow()
     {
-        public NewWindow()
-        {
-            InitializeComponent();
-        }
+        InitializeComponent();
     }
 }
