@@ -4,27 +4,28 @@ using Android.OS;
 using Application = Android.App.Application;
 using Avalonia;
 using Avalonia.Android;
+using Avalonia.ReactiveUI;
 
-namespace AvaloniaTest.Android
+namespace AvaloniaTest.Android;
+
+[Activity(Theme = "@style/MyTheme.Splash", MainLauncher = true, NoHistory = true)]
+public class SplashActivity : AvaloniaSplashActivity<App>
 {
-    [Activity(Theme = "@style/MyTheme.Splash", MainLauncher = true, NoHistory = true)]
-    public class SplashActivity : AvaloniaSplashActivity<App>
+    protected override AppBuilder CustomizeAppBuilder(AppBuilder builder)
     {
-        protected override AppBuilder CustomizeAppBuilder(AppBuilder builder)
-        {
-            return base.CustomizeAppBuilder(builder);
-        }
+        return base.CustomizeAppBuilder(builder)
+            .UseReactiveUI();
+    }
 
-        protected override void OnCreate(Bundle? savedInstanceState)
-        {
-            base.OnCreate(savedInstanceState);
-        }
+    protected override void OnCreate(Bundle? savedInstanceState)
+    {
+        base.OnCreate(savedInstanceState);
+    }
 
-        protected override void OnResume()
-        {
-            base.OnResume();
+    protected override void OnResume()
+    {
+        base.OnResume();
 
-            StartActivity(new Intent(Application.Context, typeof(MainActivity)));
-        }
+        StartActivity(new Intent(Application.Context, typeof(MainActivity)));
     }
 }
