@@ -1,15 +1,14 @@
 using UIKit;
 
-namespace AvaloniaTest.iOS
+namespace AvaloniaTest.iOS;
+
+public class Application
 {
-    public class Application
+    // This is the main entry point of the application.
+    static void Main(string[] args)
     {
-        // This is the main entry point of the application.
-        static void Main(string[] args)
-        {
-            // if you want to use a different Application Delegate class from "AppDelegate"
-            // you can specify it here.
-            UIApplication.Main(args, null, typeof(AppDelegate));
-        }
+        // if you want to use a different Application Delegate class from "AppDelegate"
+        // you can specify it here.
+        UIApplication.Main(args, null, typeof(AppDelegate));
     }
 }
