@@ -3,29 +3,28 @@ using Avalonia.Controls;
 using Avalonia.Controls.Templates;
 using AvaloniaTest.ViewModels;
 
-namespace AvaloniaTest
+namespace AvaloniaTest;
+
+public class ViewLocator : IDataTemplate
 {
-    public class ViewLocator : IDataTemplate
+    public IControl? Build(object? data)
     {
-        public IControl? Build(object? data)
+        if (data is null)
+            return null;
+
+        var name = data.GetType().FullName!.Replace("ViewModel", "View");
+        var type = Type.GetType(name);
+
+        if (type != null)
         {
-            if (data is null)
-                return null;
-
-            var name = data.GetType().FullName!.Replace("ViewModel", "View");
-            var type = Type.GetType(name);
-
-            if (type != null)
-            {
-                return (Control)Activator.CreateInstance(type)!;
-            }
-            
-            return new TextBlock { Text = name };
+            return (Control)Activator.CreateInstance(type)!;
         }
+        
+        return new TextBlock { Text = name };
+    }
 
-        public bool Match(object? data)
-        {
-            return data is ViewModelBase;
-        }
+    public bool Match(object? data)
+    {
+        return data is ViewModelBase;
     }
 }
