@@ -1,7 +1,6 @@
-﻿namespace AvaloniaTest.ViewModels
+﻿namespace AvaloniaTest.ViewModels;
+
+public class MainViewModel : ViewModelBase
 {
-    public class MainViewModel : ViewModelBase
-    {
-        public string Greeting => "Welcome to Avalonia!";
-    }
+    public string Greeting => "Welcome to Avalonia!";
 }
